+++ src/exactly_lib/test_case_utils/string_transformer/names.py
@@ -6,9 +6,11 @@ REPLACE_TRANSFORMER_NAME = 'replace'
 FILTER_TRANSFORMER_NAME = 'filter'
 TO_UPPER_CASE = 'to-upper-case'
 TO_LOWER_CASE = 'to-lower-case'
-STRIP_TRAILING_NEW_LINES = 'strip-trailing-new-lines'
-STRIP_TRAILING_SPACE = 'strip-trailing-space'
-STRIP_SPACE = 'strip-space'
+
+STRIP_SPACE = 'strip'
+STRIP_TRAILING_SPACE_OPTION_NAME = a.OptionName(long_name='trailing-space')
+STRIP_TRAILING_NEW_LINES_OPTION_NAME = a.OptionName(long_name='trailing-new-lines')
+
 TCDS_PATH_REPLACEMENT = 'replace-test-case-dirs'
 SEQUENCE_OPERATOR_NAME = '|'
 RUN_PROGRAM_TRANSFORMER_NAME = program.RUN_PROGRAM_PRIMITIVE