+++ src/exactly_lib/test_case_utils/string_transformer/parse_string_transformer.py
@@ -5,7 +5,7 @@ from exactly_lib.test_case_utils.expression.parser import GrammarParsers
 from exactly_lib.test_case_utils.string_transformer import names
 from exactly_lib.test_case_utils.string_transformer import sdvs
 from exactly_lib.test_case_utils.string_transformer.impl import filter, replace, sequence, identity, case_converters, \
-    tcds_paths_replacement, strip_trailing_new_lines, strip_trailing_space, strip_space
+    tcds_paths_replacement, strip_space
 from exactly_lib.test_case_utils.string_transformer.impl.run_program import parse as parse_run
 from exactly_lib.util.cli_syntax.elements import argument as a
 from exactly_lib.util.name_and_value import NameAndValue
@@ -53,18 +53,8 @@ GRAMMAR = grammar.Grammar(
                               case_converters.to_upper_case__documentation())
         ),
         NameAndValue(
-            names.STRIP_TRAILING_NEW_LINES,
-            grammar.Primitive(strip_trailing_new_lines.parse,
-                              strip_trailing_new_lines.SyntaxDescription())
-        ),
-        NameAndValue(
-            names.STRIP_TRAILING_SPACE,
-            grammar.Primitive(strip_trailing_space.parse,
-                              strip_trailing_space.SyntaxDescription())
-        ),
-        NameAndValue(
             names.STRIP_SPACE,
-            grammar.Primitive(strip_space.parse,
+            grammar.Primitive(strip_space.Parser().parse,
                               strip_space.SyntaxDescription())
         ),
         NameAndValue(