+++ test/exactly_lib_test/test_case_utils/string_transformers/z_package_suite.py
@@ -3,11 +3,13 @@ import unittest
 from exactly_lib_test.test_case_utils.string_transformers import \
     parse_string_transformer
 from exactly_lib_test.test_case_utils.string_transformers import tcds_paths_replacement, replace, \
-    filter, case_converters, sequence, identity, strip_trailing_new_lines, strip_trailing_space, strip_space
+    filter, case_converters, sequence, identity
 from exactly_lib_test.test_case_utils.string_transformers.models import \
     z_package_suite as models
 from exactly_lib_test.test_case_utils.string_transformers.run_program import \
     z_package_suite as run_program
+from exactly_lib_test.test_case_utils.string_transformers.strip import \
+    z_package_suite as strip
 from exactly_lib_test.test_case_utils.string_transformers.test_resources_test import \
     z_package_suite as test_resources_test
 
@@ -23,9 +25,7 @@ def suite() -> unittest.TestSuite:
     ret_val.addTest(run_program.suite())
     ret_val.addTest(parse_string_transformer.suite())
     ret_val.addTest(case_converters.suite())
-    ret_val.addTest(strip_trailing_new_lines.suite())
-    ret_val.addTest(strip_trailing_space.suite())
-    ret_val.addTest(strip_space.suite())
+    ret_val.addTest(strip.suite())
     ret_val.addTest(tcds_paths_replacement.suite())
     return ret_val